bot/requirements.txt
aiogram>=3.18.0
boto3>=1.37.11
httpx>=0.28.1
pillow>=11.1.0
pydantic-settings>=2.8.1
